Updating Upload Behavior
- Support multiple files
- Generate different names
- Use a custom field

+++ Model/Behavior/UploadBehavior.php
@@ -19,13 +19,14 @@ App::uses('Attachment', 'Attach.Model');
 
 class UploadBehavior extends ModelBehavior
 {
-    
+
     /**
     * Imagine Github URL
-    * 
+    *
     * @var string
     */
     const IMAGINE_URL = 'https://github.com/avalanche123/Imagine';
+    const SUPPORT_MULTIPLE_FILES = true;
 
     /**
      * Set what are the multiple models
@@ -473,8 +474,13 @@ class UploadBehavior extends ModelBehavior
             return;
         }
 
-        $file   = $model->generateName($type, $index);
-        $attach = $this->saveAttachment($model, $type, $file);
+        $file = $model->generateName($type, $index);
+        $userId = $uploadData['user_id'];
+        $name = $uploadData['custom_name'];
+        $description = $uploadData['description'];
+        $arrCustomData = array('user_id' => $userId, 'name' => $name, 'description' => $description);
+
+        $attach = $this->saveAttachment($model, $type, $file, $arrCustomData);
 
         if (!empty($uploadData['tmp_name'])) {
 
@@ -489,7 +495,7 @@ class UploadBehavior extends ModelBehavior
                         sprintf('The file %s is not an image', $file)
                     );
                 }
-                
+
                 //generate thumbs
                 $this->createThumbs($model, $type, $file);
             }
@@ -549,7 +555,8 @@ class UploadBehavior extends ModelBehavior
     protected function saveAttachment(
         Model $model,
         $type,
-        $filename
+        $filename,
+        $arrCustomData
     ) {
         $className = 'Attachment'. Inflector::camelize($type);
 
@@ -572,10 +579,13 @@ class UploadBehavior extends ModelBehavior
                 'foreign_key' => $model->id,
                 'filename' => basename($filename),
                 'type' => $type,
+                'user_id' => $arrCustomData['user_id'],
+                'name' => $arrCustomData['name'],
+                'description' => $arrCustomData['description'],
             ),
         );
 
-        if (!empty($attachment) && $attachment !== false) {
+        if (!empty($attachment) && $attachment !== false && !self::SUPPORT_MULTIPLE_FILES) {
             $this->deleteAllFiles($model, $attachment);
             $data[$className]['id'] = $attachment[$className]['id'];
         } else {
@@ -583,7 +593,7 @@ class UploadBehavior extends ModelBehavior
         }
 
         $model->data += $model->{$className}->save($data);
-        
+
     }
 
     /**
@@ -599,6 +609,7 @@ class UploadBehavior extends ModelBehavior
     public function generateName(Model $model, $type, $index = null)
     {
         $dir = $this->getUploadFolder($model, $type);
+        $date = new DateTime();
 
         if (is_null($index)) {
             $extension = $this->getFileExtension(
@@ -613,17 +624,19 @@ class UploadBehavior extends ModelBehavior
         }
 
         if (!is_null($index)) {
-            return $dir 
-                . $type 
-                . '_' 
-                . $index 
-                . '_' 
-                . $model->id 
-                . '.' 
+            return $dir
+                . $type
+                . '_'
+                . $index
+                . '_'
+                . $model->id
+                . '_'
+                . $date->getTimestamp()
+                . '.'
                 . $extension;
         }
 
-        return $dir . $type . '_' . $model->id  . '.' . $extension;
+        return $dir . $type . '_' . $model->id  . '_' . $date->getTimestamp() . '.' . $extension;
     }
 
     /**
